Add co-operative multitasking of PSO tasks
Each batch of PSO will now run concurrently, therefore each simulation
will run in parallel as they need wait on each other. This should mean
that to train the agents per scenario only takes as long as the slowest
scenario (2?).

Given the async/await API of Python 3.7, this should be easily extended
to allow many:many (particle swarm optimizers):(scenario training sets).

Or, at least, this is the hope. There are very likely a few bugs with
regards to sandboxing stdout into logs or the actual swapping of
execution.

Fixes #10.