Add comm & viz fixes
The comms kept completely crashing if a neighbor wasn't found (usually
for really bad reasons, but I'd rather not see it completely crash)

As it turns out, I was incorrectly looking at the neighbor's positions
and velocities! I was only looking at the x instead of x and z. (y is
always 0)