Set non-linearity to softsign -1..1
I ended up sticking with -1..1 as it is more versatile
for later fundamental changes to trying to improve generalization.

it's easy to switch to 0..1, though, just for a final test.

Also, I chose softsign over atan because it seemed to have somewhat
nicer visualizations, and it mathematically makes more sense if we're
going for a sigmoidal function with slow saturations. It was basically
made for that purpose!