Add elastic regularization
Not sure if it's correct, but it runs and would at least need parameter
tuning!
Also, the usefulness of best.tsv (renamed to loss.tsv for clarity) is
confuing. For multiset optimization the global_min value was... bizarre
and didn't correlate well with the baseline's value. I changed it to use
the same function for both, but that just makes it somewhat confusing to
interpret? I think it's easy to say that if L(c) < L(ttc), then C-TTC
is clearly better. However, if L(c) > L(ttc), that doesn't mean it's
worse, just overly complicated for what it provides...? maybe.

If all you want to know is the performance, looking at the cross_val is
probably best. There's nothing right now that immediately outputs the
sum(c)/sum(ttc), but that's not too hard to get from the cross_val.