Add median volume selection
I am entirely unsure if this would be helpful. I only implemented it
because I convinced myself that it would have a sort of regularizing
effect on the intensity of the communication, but not nearly so strong
as taking the min(sum(abs(c)) would. It appears to do that, but I think
it also introduces a fair bit of noise, considering it's based on the
distance as well. (Therefore it will select j such that it is mildly
distant from i.)